+++ crates/node-depman/tests/versions_test.rs
@@ -55,6 +55,22 @@ mod npm {
             }
         );
     }
+
+    #[test]
+    fn parses_engines() {
+        let sandbox = create_empty_sandbox();
+        let plugin = create_plugin("npm-test", sandbox.path());
+
+        assert_eq!(
+            plugin.parse_version_file(ParseVersionFileInput {
+                content: r#"{ "engines": { "npm": "1.2.3" } }"#.into(),
+                file: "package.json".into(),
+            }),
+            ParseVersionFileOutput {
+                version: Some("1.2.3".into()),
+            }
+        );
+    }
 }
 
 mod pnpm {
@@ -108,6 +124,22 @@ mod pnpm {
             }),
             ParseVersionFileOutput {
                 version: Some("latest".into()),
+            }
+        );
+    }
+
+    #[test]
+    fn parses_engines() {
+        let sandbox = create_empty_sandbox();
+        let plugin = create_plugin("pnpm-test", sandbox.path());
+
+        assert_eq!(
+            plugin.parse_version_file(ParseVersionFileInput {
+                content: r#"{ "engines": { "pnpm": "1.2.3" } }"#.into(),
+                file: "package.json".into(),
+            }),
+            ParseVersionFileOutput {
+                version: Some("1.2.3".into()),
             }
         );
     }
@@ -165,6 +197,22 @@ mod yarn {
             }),
             ParseVersionFileOutput {
                 version: Some("latest".into()),
+            }
+        );
+    }
+
+    #[test]
+    fn parses_engines() {
+        let sandbox = create_empty_sandbox();
+        let plugin = create_plugin("yarn-test", sandbox.path());
+
+        assert_eq!(
+            plugin.parse_version_file(ParseVersionFileInput {
+                content: r#"{ "engines": { "yarn": "1.2.3" } }"#.into(),
+                file: "package.json".into(),
+            }),
+            ParseVersionFileOutput {
+                version: Some("1.2.3".into()),
             }
         );
     }