+++ crates/node/tests/versions_test.rs
@@ -7,7 +7,7 @@ generate_resolve_versions_tests!("node-test", {
     "lts-gallium" => "16.20.2",
     "lts/fermium" => "14.21.3",
     "stable" => "18.17.1",
-    "node" => "20.5.1",
+    "node" => "20.6.1",
 });
 
 #[test]
@@ -47,5 +47,21 @@ fn sets_lts_aliases() {
             "argon", "boron", "carbon", "dubnium", "erbium", "fermium", "gallium", "hydrogen",
             "latest", "stable"
         ]
+    );
+}
+
+#[test]
+fn parses_engines() {
+    let sandbox = create_empty_sandbox();
+    let plugin = create_plugin("node-test", sandbox.path());
+
+    assert_eq!(
+        plugin.parse_version_file(ParseVersionFileInput {
+            content: r#"{ "engines": { "node": ">=16" } }"#.into(),
+            file: "package.json".into(),
+        }),
+        ParseVersionFileOutput {
+            version: Some(">=16".into()),
+        }
     );
 }